moved unit tests into a subfolder to make room for blackbox tests
