+++ tests/unit/common/generate-temp.js
@@ -1,0 +1,5 @@
+/* global require, process */
+const path = require('path');
+const {createDirectory} = require('../../../lib/utils');
+
+createDirectory(path.resolve(process.cwd(), 'tests/temp'));
