+++ tests/unit/common/inquirer-mock.js
