+++ tests/unit/utils/generate-config.test.js
@@ -3,7 +3,7 @@ const test = require('ava');
 const path = require('path');
 const fs = require('fs');
 const {mocki} = require('../common/inquirer-mock');
-const {generateConfigFile} = require('../../lib/utils/generate-config');
+const {generateConfigFile} = require('../../../lib/utils/generate-config');
 const {tempPath} = require('../common/config');
 
 test.serial('generateConfig - config option', async t => {
